Gateway/gatewayw/requirements-dev.txt
clang==6.0.0.2
codecov==2.0.15
coverage==4.5.3
Cython==0.29.1
docstr-coverage==1.0.3
hawkmoth==0.3
pylint==2.3.1
pytest==4.4.0
pytest-cov==2.6.1
Sphinx==1.8.5
sphinxcontrib-asyncio==0.2.0
sphinxcontrib-applehelp==1.0.1
sphinxcontrib-devhelp==1.0.1
sphinxcontrib-htmlhelp==1.0.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.2
sphinxcontrib-serializinghtml==1.1.3
sphinxcontrib-websupport==1.1.0
